+++ app/src/main/java/com/realityexpander/tasky/agenda_feature/data/repositories/eventRepository/local/entities/AttendeeEntity.kt
@@ -1,9 +1,9 @@
 package com.realityexpander.tasky.agenda_feature.data.repositories.eventRepository.local.entities
 
 import com.realityexpander.tasky.agenda_feature.data.common.serializers.ZonedDateTimeSerializer
-import com.realityexpander.tasky.agenda_feature.util.AttendeeId
-import com.realityexpander.tasky.agenda_feature.util.EventId
-import com.realityexpander.tasky.agenda_feature.util.UrlStr
+import com.realityexpander.tasky.agenda_feature.common.util.AttendeeId
+import com.realityexpander.tasky.agenda_feature.common.util.EventId
+import com.realityexpander.tasky.agenda_feature.common.util.UrlStr
 import com.realityexpander.tasky.core.util.Email
 import kotlinx.serialization.Serializable
 import java.time.ZonedDateTime
