+++ app/src/main/java/com/realityexpander/tasky/agenda_feature/data/repositories/eventRepository/remote/eventApi/DTOs/AttendeeDTO.kt
@@ -1,8 +1,8 @@
 package com.realityexpander.tasky.agenda_feature.data.repositories.eventRepository.remote.eventApi.DTOs
 
-import com.realityexpander.tasky.agenda_feature.util.AttendeeId
-import com.realityexpander.tasky.agenda_feature.util.EventId
-import com.realityexpander.tasky.agenda_feature.util.UrlStr
+import com.realityexpander.tasky.agenda_feature.common.util.AttendeeId
+import com.realityexpander.tasky.agenda_feature.common.util.EventId
+import com.realityexpander.tasky.agenda_feature.common.util.UrlStr
 import com.realityexpander.tasky.core.util.Email
 import com.realityexpander.tasky.core.util.UtcMillis
 import kotlinx.serialization.ExperimentalSerializationApi
@@ -14,7 +14,7 @@ import kotlinx.serialization.json.JsonNames
 data class AttendeeDTO(
     @JsonNames("userId") // input from json
     val id: AttendeeId,
-        val eventId: EventId? = null,
+    val eventId: EventId? = null,
 
     val email: Email,
     val fullName: String,
