+++ app/src/main/java/com/realityexpander/tasky/agenda_feature/domain/AgendaItem.kt
@@ -1,7 +1,7 @@
 package com.realityexpander.tasky.agenda_feature.domain
 
 import android.os.Parcelable
-import com.realityexpander.tasky.agenda_feature.util.PhotoId
+import com.realityexpander.tasky.agenda_feature.common.util.PhotoId
 import com.realityexpander.tasky.core.util.UserId
 import com.realityexpander.tasky.core.util.UuidStr
 import kotlinx.parcelize.Parcelize
