+++ app/src/main/java/com/realityexpander/tasky/agenda_feature/domain/AgendaSync.kt
@@ -1,8 +1,8 @@
 package com.realityexpander.tasky.agenda_feature.domain
 
-import com.realityexpander.tasky.agenda_feature.util.EventId
-import com.realityexpander.tasky.agenda_feature.util.ReminderId
-import com.realityexpander.tasky.agenda_feature.util.TaskId
+import com.realityexpander.tasky.agenda_feature.common.util.EventId
+import com.realityexpander.tasky.agenda_feature.common.util.ReminderId
+import com.realityexpander.tasky.agenda_feature.common.util.TaskId
 
 data class AgendaSync(
     val deleteEventIds: List<EventId>? = null,
